// WorkBridgeAngular/src/app/enums/notification-type.enum.ts
export enum NotificationType {
    // Jobs
    JOB_APPLIED = 'JOB_APPLIED',
    JOB_SHORTLISTED = 'JOB_SHORTLISTED',
    JOB_REJECTED = 'JOB_REJECTED',
    JOB_HIRED = 'JOB_HIRED',

    // Gig
    GIG_APPLICATION = 'GIG_APPLICATION',
    GIG_ORDER = 'GIG_ORDER',
    GIG_COMPLETED = 'GIG_COMPLETED',

    // Wallet
    DEPOSIT_SUCCESS = 'DEPOSIT_SUCCESS',
    WITHDRAW_APPROVED = 'WITHDRAW_APPROVED',
    WITHDRAW_REJECTED = 'WITHDRAW_REJECTED',

    // General
    SYSTEM = 'SYSTEM',
    ADMIN_MESSAGE = 'ADMIN_MESSAGE'
}